# swagger_clientmanagement.yml
swagger: '2.0'
info:
  description: >-
    This is a sample of server CLIENT-MANAGEMENT. This server is only for
    advisor operation.
  version: 1.0.0
  title: Swagger DM-ONLINE-BANKING
  contact:
    name: BOURGEOIS Thibault
    email: thibault.bourgeois@etu.u-pec.fr
host: 'localhost:25001'
basePath: /client-management
tags:
  - name: users
    description: User operation for advisor (cf. exam)
  - name: accounts
    description: Account operation for advisor (cf. exam)
schemes:
  - http
paths:
  /users:
    get:
      tags:
        - users
      summary: List all users
      description: ''
      produces:
        - application/json
      responses:
        '200':
          description: successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/UserDto'
        '403':
          description: FORBIDDEN
          schema:
            $ref: '#/definitions/ErrorModels'
    post:
      tags:
        - users
      summary: Create an user
      description: ''
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: userDto
          description: ''
          required: true
          schema:
            $ref: '#/definitions/UserDto'
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/UserDto'
        '403':
          description: FORBIDDEN
          schema:
            $ref: '#/definitions/ErrorModels'
  '/users/{id}':
    get:
      tags:
        - users
      summary: Get only one user
      description: ''
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: ID of user to return
          required: true
          type: integer
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/UserDto'
        '403':
          description: FORBIDDEN
          schema:
            $ref: '#/definitions/ErrorModels'
    put:
      tags:
        - users
      summary: Update an existing user
      description: ''
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: ID of user to return
          required: true
          type: integer
        - in: body
          name: userDto
          description: ''
          required: true
          schema:
            $ref: '#/definitions/UserDto'
      responses:
        '200':
          description: successful operation
        '403':
          description: FORBIDDEN
          schema:
            $ref: '#/definitions/ErrorModels'
  /accounts:
    post:
      tags:
        - accounts
      summary: Create an new account for user
      description: ''
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: accountDto
          description: ''
          required: true
          schema:
            $ref: '#/definitions/AccountDto'
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/AccountDto'
        '403':
          description: FORBIDDEN
          schema:
            $ref: '#/definitions/ErrorModels'
  '/accounts/users/{id}':
    get:
      tags:
        - accounts
      summary: Get list accounts for one user
      description: ''
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: ID of user to get accounts
          required: true
          type: integer
      responses:
        '200':
          description: successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/AccountDto'
        '403':
          description: FORBIDDEN
          schema:
            $ref: '#/definitions/ErrorModels'
definitions:
  UserDto:
    type: object
    properties:
      id:
        type: integer
        format: int64
        example: 1
      lastName:
        type: string
        example: BOURGEOIS
      firstName:
        type: string
        example: Thibault
      address:
        type: string
        example: 26 rue Chateaubriand
      phone:
        type: string
        example: 0629516898
      birth:
        type: string
        format: date
        example: '2017-07-06'
  AccountDto:
    type: object
    properties:
      id:
        type: integer
        format: int64
        example: 1
      userDto:
        $ref: '#/definitions/UserDto'
      accountType:
        type: string
        example: Compte courant
      solde:
        type: integer
        format: double
        example: 123.45
  ErrorModels:
    type: object
    properties:
      Message:
        type: string
        example: Aucune resources